// mock/pay/typings/data.d.ts
// Generated by https://quicktype.io

export interface FakeLists {
  id: number
  createStart: string
  price: string
  payPrice: string
  status: number
  statusChangeTime: string
  grouperId: number
  grouperName: string
  company: number
  companyName: string
  orderId: number
  orderTitle: string
  receiverName: string
  receiverPhone: string
  operator: string
  isSuccess: boolean
}